01langchain_model/requirements.txt
langchain
langchain-openai
langchain-core
openai
langchain-anthropic
langchain-google-genai
google-generativeai
langchain-huggingface
transformers
huggingface-hub
python-dotenv
numpy
scikit-learn

